Move `wallet_change_notifier` to `controller`
